Amber's version of data extraction
This is my version of the data extraction pipeline. It will generate a df for a defined session with all of the features/columns that we want except for wheel (still working on that). So, this is one version that is completed. HOWEVER, we are going to finish and use Mariana's version of this. We can use this as a second option if we run into problems.